Moved default test for request into separate files.
