+++ src/test/request/defaults.http.IncomingMessage.test.ts
@@ -1,0 +1,117 @@
+// Tested Module
+import getMockReq from '../../request/request'
+
+describe('request - Defaults from "http - IncomingMessage" (accepts no arguments and return default values)', () => {
+  test('req.aborted is a boolean', () => {
+    const req = getMockReq()
+
+    expect(req.aborted).toBeDefined()
+    expect(req.aborted).toBe(false)
+  })
+
+  test('req.httpVersion is an empty string', () => {
+    const req = getMockReq()
+
+    expect(req.httpVersion).toBeDefined()
+    expect(req.httpVersion).toBe('')
+  })
+
+  test('req.httpVersionMajor is a number', () => {
+    const req = getMockReq()
+
+    expect(req.httpVersionMajor).toBeDefined()
+    expect(req.httpVersionMajor).toBe(0)
+  })
+
+  test('req.httpVersionMinor is a number', () => {
+    const req = getMockReq()
+
+    expect(req.httpVersionMinor).toBeDefined()
+    expect(req.httpVersionMinor).toBe(0)
+  })
+
+  test('req.complete is a boolean', () => {
+    const req = getMockReq()
+
+    expect(req.complete).toBeDefined()
+    expect(req.complete).toBe(false)
+  })
+
+  test('req.connection is an empty object', () => {
+    const req = getMockReq()
+
+    expect(req.connection).toBeDefined()
+    expect(req.connection).toBeInstanceOf(Object)
+    expect(Object.keys(req.connection).length).toBe(0)
+  })
+
+  test('req.socket is an empty object', () => {
+    const req = getMockReq()
+
+    expect(req.socket).toBeDefined()
+    expect(req.socket).toBeInstanceOf(Object)
+    expect(Object.keys(req.socket).length).toBe(0)
+  })
+
+  test('req.headers is an empty object', () => {
+    const req = getMockReq()
+
+    expect(req.headers).toBeDefined()
+    expect(req.headers).toBeInstanceOf(Object)
+    expect(Object.keys(req.headers).length).toBe(0)
+  })
+
+  test('req.rawHeaders is an empty array', () => {
+    const req = getMockReq()
+
+    expect(req.rawHeaders).toBeDefined()
+    expect(Array.isArray(req.rawHeaders)).toBe(true)
+    expect(req.rawHeaders.length).toBe(0)
+  })
+
+  test('req.trailers is an empty object', () => {
+    const req = getMockReq()
+
+    expect(req.trailers).toBeDefined()
+    expect(req.trailers).toBeInstanceOf(Object)
+    expect(Object.keys(req.trailers).length).toBe(0)
+  })
+
+  test('req.rawTrailers is an empty array', () => {
+    const req = getMockReq()
+
+    expect(req.rawTrailers).toBeDefined()
+    expect(Array.isArray(req.rawTrailers)).toBe(true)
+    expect(req.rawTrailers.length).toBe(0)
+  })
+
+  test('req.setTimeout is a mocked function', () => {
+    const req = getMockReq()
+
+    expect(req.setTimeout).toBeDefined()
+    expect(typeof req.setTimeout).toBe('function')
+    expect((req.setTimeout as jest.Mock).getMockName()).toBe('setTimeout mock default')
+  })
+
+  test('req.statusCode is a number', () => {
+    const req = getMockReq()
+
+    expect(req.statusCode).toBeDefined()
+    expect(req.statusCode).toBe(0)
+  })
+
+  test('req.statusMessage is an empty string', () => {
+    const req = getMockReq()
+
+    expect(req.statusMessage).toBeDefined()
+    expect(req.statusMessage).toBe('')
+  })
+
+  test('req.destroy is a mocked function', () => {
+    const req = getMockReq()
+
+    expect(req.destroy).toBeDefined()
+    expect(typeof req.destroy).toBe('function')
+    expect((req.destroy as jest.Mock).getMockName()).toBe('destroy mock default')
+  })
+})
